Move view controller helpers to Stripe Core.
